Made return type of ncclTreeThreshold() int64_t
Otherwise we get an overflow and -ve numbers at 2GB+

+++ src/include/core.h
@@ -404,7 +404,7 @@ struct ncclComm {
 #endif // end PROFAPI
 
 int ncclCudaCompCap();
-int ncclTreeThreshold();
+int64_t ncclTreeThreshold();
 
 static __inline__ int ncclTypeSize(ncclDataType_t type) {
   switch (type) {
